Direction 'B' done
Now, the enemies can move to right for X times, then move to left for X times.
This happens in 240 line to 253.

For this, we created a flag to show if direction is 'B' in the level. If it is, we change the level's direction per X time, between 'R' and 'L'.